try to kill warning

+++ db/extsort.cpp
@@ -221,7 +221,7 @@ namespace mongo {
         unsigned long long length;
         _buf = (char*)_file.map( file.c_str() , length , MemoryMappedFile::SEQUENTIAL );
         massert( 10308 ,  "mmap failed" , _buf );
-        assert( length == file_size( file ) );
+        assert( length == (unsigned long long) file_size( file ) );
         _end = _buf + length;
     }
     BSONObjExternalSorter::FileIterator::~FileIterator() {}
